requirements.txt
canvasapi
openai
python-dotenv
pandas
